removed separate mutation for updating settings on subspaces
